Abstract away the driver ui into it's own class called DriverWidget
Add a new class called DriverWidget to which you can pass the label text,
the package map and the QApt backed from which it will draw the ui. This class
also returns the currently selected package to the main KCM

Reuse the KMessageWidget by adding it to the ui file and using it everywhere.